separate search function into different file
