+++ train-tracker/src/App.js
@@ -6,6 +6,7 @@ import Map from './Map';
 import TrainList from './TrainList';
 import Search from './Search';
 import TrainPopup from './TrainPopup';
+import {sortTrains} from './functionality/app.js'
 
 import { IoClose } from "react-icons/io5";
 
@@ -49,32 +50,8 @@ function App() {
     function handleFromStation(e){ setFromStation(e.target.value); }
     function handleToStation(e){ setToStation(e.target.value); }
 
-    const sortTrains = () => {
-        let trains = allTrains;
-        if (selectedNumber > 0){
-            trains = trains.filter(t => t.number === selectedNumber)
-        } if (selectedRoute){
-            trains = trains.filter(t => t.routeName === selectedRoute)
-        } if (selectedStation){
-            trains = trains.filter((t,index) => (t.stations.findIndex((station) => station.stationCode === selectedStation) !== -1));
-            if (upcoming){
-                trains = trains.filter((t, index) => {
-                    let station = t.stations.find((station) => station.stationCode === selectedStation);
-                    if (station.stationCode === t.from && station.hasDeparted){
-                        return;
-                    }
-                    if (!station.hasArrived || !station.hasDeparted){
-                        return t;
-                    }
-                })
-            }
-        } if (fromStation && toStation){
-            trains = trains.filter((t, index) =>{
-                return t.stations.findIndex((station) => station.stationCode === fromStation) < t.stations.findIndex((station) => station.stationCode === toStation);
-            })
-        }
-        // sort results by number
-        trains.sort((a,b) => a.number - b.number);
+    const searchTrains = () => {
+        let trains = sortTrains(allTrains, selectedNumber, selectedRoute, selectedStation, upcoming, fromStation, toStation);
         setCurrentTrains(trains);
     }
 
@@ -134,7 +111,7 @@ function App() {
                 route = {selectedRoute} setRoute = {handleRoute} routes = {getRouteOptions()}
                 station = {selectedStation} setStation = {handleStation} stations = {getStationOptions()}
                 upcoming = {upcoming} setUpcoming = {handleUpcoming}
-                searchFun = {sortTrains} clearFun = {clearSearch}
+                searchFun = {searchTrains} clearFun = {clearSearch}
                 fromStation = {fromStation} setFromStation = {handleFromStation}
                 toStation = {toStation} setToStation = {handleToStation}
               />
